Moving i18n files to .mjs
