+++ packages/core/src/i18n/translations/en.mjs
