+++ packages/core/src/i18n/translations/tl.mjs
