+++ packages/web-components/src/index.ts
@@ -1,6 +1,6 @@
-import '../../core/src/i18n/i18n-setup';
+import '../../core/src/i18n/i18n-setup.mjs';
 
-export { i18next } from '../../core/src/i18n/i18n-setup';
+export { i18next } from '../../core/src/i18n/i18n-setup.mjs';
 
 export { Components, JSX } from './components';
